+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools" >
 
     <application
         android:allowBackup="true"
@@ -11,9 +11,14 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.ANDANDROID"
-        tools:targetApi="31">
+        tools:targetApi="31" >
         <activity
             android:name=".LogInActivity"
+            android:exported="false"
+            android:label="@string/title_activity_log_in"
+            android:theme="@style/Theme.ANDANDROID" />
+        <activity
+            android:name=".SignUpActivity"
             android:exported="false"
             android:label="@string/title_activity_log_in"
             android:theme="@style/Theme.ANDANDROID" />
