+++ app/src/main/res/drawable/ic_search.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M16.927,17.04L20.4,20.4M19.28,11.44C19.28,15.77 15.77,19.28 11.44,19.28C7.11,19.28 3.6,15.77 3.6,11.44C3.6,7.11 7.11,3.6 11.44,3.6C15.77,3.6 19.28,7.11 19.28,11.44Z"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"
+      android:strokeLineCap="round"/>
+</vector>
