+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,9 @@
 <resources>
     <string name="app_name">AND-ANDROID</string>
     <string name="title_activity_log_in">LogInActivity</string>
+    <string name="title_activity_main">MainActivity</string>
+
+    <string name="home">홈</string>
+    <string name="search">검색</string>
+    <string name="my">마이페이지</string>
 </resources>
